requirements.txt
streamlit==1.30.0
roboflow==1.1.16
Pillow==10.2.0
